Move generated CrossRegionS3Client to a separate module.
