+++ spring-cloud-aws-dependencies/pom.xml
@@ -88,6 +88,12 @@
 
 			<dependency>
 				<groupId>io.awspring.cloud</groupId>
+				<artifactId>spring-cloud-aws-s3-cross-region-client</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+
+			<dependency>
+				<groupId>io.awspring.cloud</groupId>
 				<artifactId>spring-cloud-starter-aws-parameter-store</artifactId>
 				<version>${project.version}</version>
 			</dependency>
