+++ spring-cloud-aws-s3-parent/spring-cloud-aws-s3/pom.xml
@@ -43,59 +43,4 @@
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
-
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.codehaus.mojo</groupId>
-				<artifactId>exec-maven-plugin</artifactId>
-				<version>3.0.0</version>
-				<dependencies>
-					<dependency>
-						<groupId>io.awspring.cloud</groupId>
-						<artifactId>spring-cloud-aws-s3-codegen</artifactId>
-						<version>${project.version}</version>
-					</dependency>
-				</dependencies>
-				<executions>
-					<execution>
-						<id>generate-javaparser-core</id>
-						<phase>generate-sources</phase>
-						<goals>
-							<goal>java</goal>
-						</goals>
-					</execution>
-				</executions>
-				<configuration>
-					<includeProjectDependencies>true</includeProjectDependencies>
-					<includePluginDependencies>true</includePluginDependencies>
-					<mainClass>io.awspring.cloud.s3.codegen.CrossRegionS3ClientGenerator</mainClass>
-					<arguments>
-						<argument>${project.basedir}</argument>
-					</arguments>
-				</configuration>
-			</plugin>
-			<!-- disable checkstyle as generated code does not match required style -->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-checkstyle-plugin</artifactId>
-				<executions>
-					<execution>
-						<id>checkstyle-validation</id>
-						<phase>none</phase>
-					</execution>
-				</executions>
-			</plugin>
-			<!-- disable spring-javaformat as generated code does not match required formatting -->
-			<plugin>
-				<groupId>io.spring.javaformat</groupId>
-				<artifactId>spring-javaformat-maven-plugin</artifactId>
-				<executions>
-					<execution>
-						<phase>none</phase>
-					</execution>
-				</executions>
-			</plugin>
-		</plugins>
-	</build>
 </project>
